+++ crates/relayer-types/src/clients/ics08_cardano/consensus_state.rs
@@ -4,8 +4,8 @@ use serde_derive::{Deserialize, Serialize};
 use ibc_proto::google::protobuf::Any;
 use ibc_proto::Protobuf;
 
-use crate::clients::ics2000_mithril::error::Error;
-use crate::clients::ics2000_mithril::raw as raw;
+use crate::clients::ics08_cardano::error::Error;
+use crate::clients::ics08_cardano::raw as raw;
 use crate::core::ics02_client::client_type::ClientType;
 use crate::core::ics02_client::consensus_state::ConsensusState as Ics2ConsensusState;
 use crate::core::ics02_client::error::Error as Ics02Error;
