+++ crates/relayer-types/src/clients/ics08_cardano/header.rs
@@ -4,8 +4,8 @@ use ibc_proto::Protobuf;
 use prost::Message;
 use serde_derive::{Deserialize, Serialize};
 
-use crate::clients::ics2000_mithril::error::Error;
-use crate::clients::ics2000_mithril::raw as raw;
+use crate::clients::ics08_cardano::error::Error;
+use crate::clients::ics08_cardano::raw as raw;
 use crate::core::ics02_client::client_type::ClientType;
 use crate::core::ics02_client::error::Error as Ics02Error;
 use crate::timestamp::Timestamp;
