--- crates/relayer-types/src/clients/ics2000_mithril/mod.rs
@@ -1,14 +1,0 @@
-//! ICS-2000: Cardano Mithril Client
-//!
-//! This module contains the types used by the Cosmos-sidechain Mithril light client
-//! (`08-cardano`), as defined in `ibc.lightclients.mithril.v1`.
-
-pub mod client_state;
-pub mod consensus_state;
-pub mod error;
-pub mod header;
-pub mod raw;
-
-pub use client_state::ClientState;
-pub use consensus_state::ConsensusState;
-pub use header::Header;
