+++ crates/relayer-types/src/clients/mod.rs
@@ -1,4 +1,4 @@
 //! Implementations of client verification algorithms for specific types of chains.
 
 pub mod ics07_tendermint;
-pub mod ics2000_mithril;
+pub mod ics08_cardano;
